+++ packages/picasso-lab/src/TimePicker/TimePicker.tsx
@@ -83,6 +83,7 @@ export const TimePicker = (props: Props) => {
       />
     )
   }
+
   return (
     <Input
       type='time'