+++ packages/picasso-lab/src/TypographyOverflow/TypographyOverflow.tsx
@@ -1,4 +1,4 @@
-import React, { ReactText } from 'react'
+import React, { ReactNode } from 'react'
 import { BaseProps } from '@toptal/picasso-shared'
 import { Tooltip, Typography, TypographyProps } from '@toptal/picasso'
 
@@ -6,10 +6,17 @@ import Ellipsis from '../Ellipsis'
 
 export interface Props extends BaseProps, TypographyProps {
   /** A typography which can possibly overflow */
-  children: ReactText
+  children?: ReactNode
+  /** A content to show in tooltip when typography overflows. By default, TypographyOverflow's children are used. */
+  tooltipContent?: ReactNode
 }
 
-export const TypographyOverflow = ({ children, noWrap, ...rest }: Props) => {
+export const TypographyOverflow = ({
+  children,
+  tooltipContent,
+  noWrap,
+  ...rest
+}: Props) => {
   const typography = (
     // eslint-disable-next-line react/jsx-props-no-spreading
     <Typography {...rest} noWrap={noWrap}>
@@ -20,7 +27,11 @@ export const TypographyOverflow = ({ children, noWrap, ...rest }: Props) => {
   return (
     <Ellipsis
       renderWhenEllipsis={child => (
-        <Tooltip content={children} placement='top' interactive>
+        <Tooltip
+          content={tooltipContent ?? children}
+          placement='top'
+          interactive
+        >
           {child}
         </Tooltip>
       )}