+++ packages/picasso/src/Button/story/Disabled.example.jsx
@@ -1,5 +1,6 @@
 import React from 'react'
 import { Button } from '@toptal/picasso'
+import { Trash24 } from '@toptal/picasso/Icon'
 
 const Example = () => (
   <div>
@@ -7,6 +8,7 @@ const Example = () => (
     <Button disabled variant='secondary-red'>
       Secondary Disabled
     </Button>
+    <Button disabled variant='transparent-green' icon={<Trash24 />} />
   </div>
 )